+++ joe/utf8.c
@@ -364,7 +364,10 @@ void joe_locale()
 	if ((t=zrchr(s,'.')))
 		*t = 0;
 
-	locale_msgs = s;
+	if (!zicmp(s, USTR "C") || !zicmp(s, USTR "POSIX"))
+		locale_msgs = USTR "en_US"; /* Because aspell has no default */
+	else
+		locale_msgs = s;
 
 	s=(unsigned char *)getenv("LC_ALL");
 	if (!s || !*s) {